requirements.txt
# =====================================================
# Legislative Monitor - Granite Edition
# Python 3.10 or 3.11 REQUIRED
# Optimized for AMD hardware with ROCm support
# =====================================================

# Note: PyTorch is installed separately by Dockerfile to handle ROCm/CUDA/CPU

# Primary Transcription - IBM Granite Speech
transformers>=4.52.4
peft

# Fallback Transcription - Whisper
faster-whisper>=1.0.3

# Speaker Diarization - SpeechBrain (6.7x faster than Pyannote)
speechbrain>=1.0.0

# Hugging Face Hub
huggingface_hub>=0.20.2
datasets>=2.16.1

# Audio Processing
av>=13.0.0  # PyAV - must be first to ensure FFmpeg 7.x compatibility
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
torchaudio  # Installed with PyTorch

# Scientific Computing
numpy>=1.24.3
scipy>=1.11.4
scikit-learn>=1.3.2  # Required for MeanShift clustering in diarization

# Video Download
yt-dlp==2024.12.23

# Web Scraping & HTTP
beautifulsoup4==4.12.3
requests==2.31.0
httpx>=0.27.0  # Async HTTP client for webhook
lxml==5.1.0

# Document Generation
python-docx==1.1.0

# Task Scheduling
schedule==1.2.1

# Environment & Config
python-dotenv==1.0.1

# Progress Bars
tqdm==4.66.1

# Audio Event Classification (for sound detection)
panns-inference==0.1.0

# WebDAV Client for Seafile
webdav4==0.9.8

# SFTP Client
paramiko>=3.0.0

# Additional Utilities
Pillow>=10.2.0
